Refactor domain contracts into per-entity files
